+++ resolwe_bio/tests/processes/test_reads_filtering.py
@@ -284,7 +284,7 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
     def test_bamclipper(self):
         species = 'Homo sapiens'
         build = 'fake_genome_RSEM'
-        align_input = './bamclipper/input/STK11.bam'
+        align_input = './bamclipper/input/TP53.bam'
 
         with self.preparation_stage():
             bam = self.prepare_bam(
@@ -293,7 +293,7 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
                 build=build
             )
 
-            inputs_bedpe = {'src': './bamclipper/input/STK11.bedpe',
+            inputs_bedpe = {'src': './bamclipper/input/TP53.bedpe',
                             'species': species, 'build': build}
             bedpe = self.run_process('upload-bedpe', inputs_bedpe)
 
@@ -308,8 +308,8 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         inputs_bamclipper = {'alignment': bam.id, 'bedpe': bedpe.id}
         clipped = self.run_process('bamclipper', inputs_bamclipper)
 
-        self.assertFile(clipped, 'stats', './bamclipper/output/STK11.primerclipped.bam_stats.txt')
-        self.assertFile(clipped, 'bigwig', './bamclipper/output/STK11.primerclipped.bw')
+        self.assertFile(clipped, 'stats', './bamclipper/output/TP53.primerclipped.bam_stats.txt')
+        self.assertFile(clipped, 'bigwig', './bamclipper/output/TP53.primerclipped.bw')
         self.assertFields(clipped, 'species', species)
         self.assertFields(clipped, 'build', build)
 
@@ -317,7 +317,7 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
     def test_markduplicates(self):
         species = 'Homo sapiens'
         build = 'custombuild'
-        primerclipped = './bamclipper/output/STK11.primerclipped.bam'
+        primerclipped = './bamclipper/output/TP53.primerclipped.bam'
 
         with self.preparation_stage():
             bam = self.prepare_bam(
@@ -325,7 +325,7 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
                 species=species,
                 build=build)
 
-        # Teste if skipped. Input bam should always equal output bam.
+        # Test if skipped. Input bam should always equal output bam.
         md_inputs = {'bam': bam.id, 'skip': True}
         skipped_md = self.run_process('markduplicates', md_inputs)
         self.assertFile(skipped_md, 'bam', primerclipped)
@@ -339,9 +339,9 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
 
         self.assertFileExists(removed_md, 'bam')
         self.assertFileExists(removed_md, 'bai')
-        self.assertFile(removed_md, 'stats', './markduplicate/output/STK11.primerclipped.markduplicates.bam_stats.txt')
-        self.assertFile(removed_md, 'bigwig', './markduplicate/output/STK11.primerclipped.markduplicates.bw')
-        self.assertFile(removed_md, 'metrics_file', './markduplicate/output/STK11.primerclipped_metrics.txt',
+        self.assertFile(removed_md, 'stats', './markduplicate/output/TP53.primerclipped.markduplicates.bam_stats.txt')
+        self.assertFile(removed_md, 'bigwig', './markduplicate/output/TP53.primerclipped.markduplicates.bw')
+        self.assertFile(removed_md, 'metrics_file', './markduplicate/output/TP53.primerclipped_metrics.txt',
                         file_filter=filter_startedon)
         self.assertFields(removed_md, 'species', species)
         self.assertFields(removed_md, 'build', build)
@@ -355,33 +355,24 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
                 # Based on b37 genome, chromosome 19 has been cut from beginning up to position 1207173.
                 # This includes an exon of STK11. Cutting from the start of the chromosome was done so that
                 # there is no need to shift any subsequent bed and vcf files.
-                'src': './bqsr/input/hs_b37_chr19_upto_stk11.fasta.gz',
+                'src': './bqsr/input/hs_b37_chr17_upto_TP53.fasta.gz',
                 'species': species,
                 'build': build
             }
             input_bam = {
-                'src': './markduplicate/output/STK11.primerclipped.markduplicates.bam',
+                'src': './markduplicate/output/TP53.primerclipped.markduplicates.bam',
                 'species': species,
                 'build': build
             }
 
             ks_dbsnp = []
-            # dbSNP vcf has been cut to only specific region using VCFtools - 0.1.15. dbSNP database version is 138.
-            # vcftools --gzvcf dbsnp_138.b37.vcf.gz \
-            # --chr 19 \
-            # --from-bp 1206985 \
-            # --to-bp 1207173 \
-            # --recode --recode-INFO-all \
-            # --stdout > dbsnp_STK11.vcf
-            # Header has been fixed by removing all but contig 19. Length has also been changed
-            # to 1207200.
-            for i in ['./bqsr/input/dbsnp_STK11.vcf.gz']:  # add more files if needed
+            for i in ['./bqsr/input/dbsnp_TP53.vcf.gz']:  # add more files if needed
                 ks_dbsnp.append(
                     self.run_process('upload-variants-vcf', {'src': i, 'species': species, 'build': build})
                 )
 
             intervals = self.run_process('upload-bed', {
-                'src': './bqsr/input/STK11.bed',
+                'src': './bqsr/input/TP53.bed',
                 'species': species,
                 'build': build})
 
@@ -398,10 +389,10 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
 
             self.assertFileExists(bqsr, 'bam')
             self.assertFileExists(bqsr, 'bai')
-            self.assertFile(bqsr, 'stats', './bqsr/output/STK11.primerclipped.markduplicates.bam_stats.txt')
-            self.assertFile(bqsr, 'bigwig', './bqsr/output/STK11.primerclipped.markduplicates.bw')
+            self.assertFile(bqsr, 'stats', './bqsr/output/TP53.primerclipped.markduplicates.bam_stats.txt')
+            self.assertFile(bqsr, 'bigwig', './bqsr/output/TP53.primerclipped.markduplicates.bw')
             self.assertFile(bqsr, 'recal_table',
-                            './bqsr/output/STK11.primerclipped.markduplicates_recalibration.table')
+                            './bqsr/output/TP53.primerclipped.markduplicates_recalibration.table')
             self.assertFields(bqsr, 'species', species)
             self.assertFields(bqsr, 'build', build)