+++ docs/CHANGELOG.rst
@@ -5,6 +5,15 @@ Change Log
 All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+==========
+Unreleased
+==========
+
+Changed
+-------
+- Test files improved for ``workflow-wes``, ``bamclipper``,
+  ``markduplicates`` and ``bqsr``
+
 
 ===================
 23.0.0 - 2019-09-17